Refactor recipe for cfgfile.
