--- recipes/cfgfile/0.2.8.2.1/conandata.yml
@@ -1,4 +1,0 @@
-sources:
-  "0.2.8.2.1":
-    url: "https://github.com/igormironchik/cfgfile/releases/download/0.2.8.2.1/cfgfile-0.2.8.2.1-full.tar.gz"
-    sha256: "6b0fe28323904ceb4567c5f3d450ab6eec6cca5eebad7ad0b77b1967f9e2772d"
