--- recipes/cfgfile/0.2.8.2.1/test_package/conanfile.py
@@ -1,15 +1,0 @@
-from conans import ConanFile, CMake, tools
-import os
-
-class CfgfileTestConan(ConanFile):
-    generators = "cmake"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if not tools.cross_building(self.settings):
-            bin_path = os.path.join("bin", "cfgfile.test")
-            self.run(bin_path, run_environment=True)
