--- recipes/cfgfile/0.2.9.0/conandata.yml
@@ -1,4 +1,0 @@
-sources:
-  "0.2.9.0":
-    url: "https://github.com/igormironchik/cfgfile/archive/refs/tags/0.2.9.0.tar.gz"
-    sha256: "4f42899849e87779089aab1db519ad7977c65cc40315c15515d7e29b0a3088c8"
