--- recipes/cfgfile/0.2.9.0/test_package/test.cfg
@@ -1,3 +1,0 @@
-{Session
-    {project 123}
-}
