--- recipes/cfgfile/0.2.9.1/conandata.yml
@@ -1,4 +1,0 @@
-sources:
-  "0.2.9.1":
-    url: "https://github.com/igormironchik/cfgfile/archive/refs/tags/0.2.9.1.tar.gz"
-    sha256: "b5c8aa2ab148404325e817dc059dd6af81e96d7d831f137c93fe9dd79bbe3c7b"
