--- recipes/cfgfile/0.2.9.1/test_package/session_cfg.conf
@@ -1,19 +1,0 @@
-
-{forGeneration SESSION_CFG_HPP
-
-    {namespace Cfg
-
-        |#
-            Session.
-        #|
-        {class Session
-            {tagScalar
-                {valueType std::string}
-                {name project}
-                {required}
-            }
-        }
-
-    } || namespace Cfg
-
-}
