+++ recipes/cfgfile/all/CMakeLists.txt
