+++ recipes/cfgfile/all/conanfile.py
@@ -1,6 +1,7 @@
 from conans import ConanFile, CMake, tools
 from conans.errors import ConanInvalidConfiguration
 import os
+import textwrap
 
 
 class CfgfileConan(ConanFile):
@@ -10,7 +11,7 @@ class CfgfileConan(ConanFile):
     license = "MIT"
     description = "Header-only library for reading/saving configuration files with schema defined in sources."
     exports_sources = "CMakeLists.txt"
-    generators = "cmake"
+    generators = "cmake", "cmake_find_package"
     topics = ("conan", "cfgfile", "configuration")
     settings = "os", "arch", "compiler", "build_type"
 
@@ -39,6 +40,8 @@ class CfgfileConan(ConanFile):
         self._cmake = CMake(self)
         self._cmake.definitions["BUILD_EXAMPLES"] = False
         self._cmake.definitions["BUILD_TESTS"] = False
+        if tools.Version(self.version) >= "0.2.9":
+            self._cmake.definitions["USE_INTERNAL_ARGS_PARSER"] = False
         self._cmake.configure(build_folder=self._build_subfolder)
         return self._cmake
 
@@ -54,6 +57,12 @@ class CfgfileConan(ConanFile):
         version = tools.Version(self.settings.compiler.version)
         if version < self._compilers_minimum_version[compiler]:
             raise ConanInvalidConfiguration("cfgfile requires a compiler that supports at least C++14")
+
+    def requirements(self):
+        if self.version == "0.2.9.1":
+            self.requires("args-parser/6.2.0.1")
+        elif self.version == "0.2.9.0":
+            self.requires("args-parser/6.0.1.0")
 
     def build(self):
         cmake = self._configure_cmake()
@@ -72,6 +81,9 @@ class CfgfileConan(ConanFile):
         bin_path = os.path.join(self.package_folder, "bin")
         self.output.info("Appending PATH env var with : {}".format(bin_path))
         self.env_info.PATH.append(bin_path)
+        self.cpp_info.names["cmake_find_package"] = "cfgfile"
+        self.cpp_info.names["cmake_find_package_multi"] = "cfgfile"
+        self.cpp_info.includedirs.append(os.path.join("include", "cfgfile"))
 
     def package_id(self):
         del self.info.settings.compiler
