+++ recipes/cfgfile/all/test_package/example.cpp
