+++ recipes/cfgfile/config.yml
@@ -1,0 +1,7 @@
+versions:
+  "0.2.9.1":
+    folder: all
+  "0.2.9.0":
+    folder: all
+  "0.2.8.2.1":
+    folder: all
